First milestone reached
Game now has a board that can be drawn to the screen. A problem that
ocurred where that I drew the outlines of the tiles first then filled
the tile wich overdrawed the tile so no outlines where shown. Now when I
switched the order of the method calls it works.