Added constructor to abstractRobot. Added hitPoints field to abstract robot. Added class testRobot to create a robot for testing.
Implemented move method in abstractRobot, robots can now be moved across the board.
Implemented temporary draw method in abstractRobot, robots are now represented by a blue square and a string showing the direction.
Implemented a moveForwardOne method in testRobot to move the robot one tile in the direction it is facing.
Implemented turnLeft and turnRight methods in testRobot to turn robot 90 degrees to the left or right.

Need method(s) in tiles to be able to ad a robot to a tile
Need a way to add robots to the game(From the "Game" class?)